+++ packages/cli/src/commands/app/create.ts
@@ -1,6 +1,7 @@
 import { ReadStream } from 'fs';
 
 import { logger } from '@appsemble/node-utils';
+import { AppVisibility } from '@appsemble/types';
 import fg from 'fast-glob';
 import normalizePath from 'normalize-path';
 import { Argv } from 'yargs';
@@ -16,11 +17,14 @@ interface CreateAppArguments extends BaseArguments {
   maskableIcon: NodeJS.ReadStream | ReadStream;
   paths: string[];
   organization: string;
-  private: boolean;
   template: boolean;
   dryRun: boolean;
   resources: boolean;
   modifyContext: boolean;
+  visibility: AppVisibility;
+  sentryDsn: string;
+  sentryEnvironment: string;
+  googleAnalyticsId: string;
 }
 
 export const command = 'create <paths...>';
@@ -50,50 +54,42 @@ export function builder(yargs: Argv): Argv {
         'The maskable icon to upload. By default "maskable-icon.png" in the app directory is used.',
       coerce: coerceFile,
     })
-    .option('private', {
-      describe: 'Whether the app should be marked as private.',
-      default: true,
-      type: 'boolean',
+    .option('visibility', {
+      describe: 'Visibility of the app in the public app store.',
+      default: 'unlisted',
+      choices: ['public', 'unlisted', 'private'],
     })
     .option('template', {
       describe: 'Whether the app should be marked as a template.',
-      default: false,
       type: 'boolean',
     })
     .option('dry-run', {
       describe: 'Whether the API should be called to run without actually creating the app.',
-      default: false,
       type: 'boolean',
     })
     .option('resources', {
       describe:
         'Whether the resources from the `resources` directory should be created after creating the app. The names of subdirectories are used as the name of the resource, otherwise the names of top level resource .json files are used instead.',
-      default: false,
       type: 'boolean',
     })
     .option('modify-context', {
       describe:
         'If the app context is specified, modify it for the current context to include the id of the created app.',
-      default: false,
       type: 'boolean',
+    })
+    .option('google-analytics-id', {
+      describe: 'The ID for Google Analytics for the app.',
+    })
+    .option('sentry-dsn', {
+      describe: 'The custom Sentry DSN for the app.',
+    })
+    .option('sentry-environment', {
+      describe: 'The environment for the custom Sentry DSN for the app.',
+      implies: ['sentry-dsn'],
     });
 }
 
-export async function handler({
-  clientCredentials,
-  context,
-  dryRun,
-  icon,
-  iconBackground,
-  maskableIcon,
-  modifyContext,
-  organization,
-  paths,
-  private: isPrivate,
-  remote,
-  resources,
-  template,
-}: CreateAppArguments): Promise<void> {
+export async function handler({ paths, ...args }: CreateAppArguments): Promise<void> {
   const normalizedPaths = paths.map((path) => normalizePath(path));
   const directories = await fg(normalizedPaths, { absolute: true, onlyDirectories: true });
 
@@ -101,19 +97,8 @@ export async function handler({
   for (const dir of directories) {
     logger.info('');
     await createApp({
-      clientCredentials,
-      context,
-      organization,
+      ...args,
       path: dir,
-      icon,
-      iconBackground,
-      maskableIcon,
-      private: isPrivate,
-      remote,
-      template,
-      dryRun,
-      resources,
-      modifyContext,
     });
   }
 }