+++ packages/studio/src/pages/apps/app/settings/types.ts
@@ -1,11 +1,16 @@
+import { AppVisibility } from 'types/src/app';
+
 export interface FormValues {
   googleAnalyticsID: string;
+  sentryDsn: string;
+  sentryEnvironment: string;
   icon: File | string;
   maskableIcon?: File;
   iconBackground: string;
   path: string;
-  private: boolean;
+  visibility: AppVisibility;
   domain: string;
   locked: boolean;
   longDescription: string;
+  showAppDefinition: boolean;
 }